server/requirements.txt
photoshop_connection==0.1.2
requests==2.23.0
opencv_python==4.1.0.25
numpy==1.16.4
Django==3.0.2
pyscreenshot==1.0
djangorestframework==3.11.0
Pillow==7.1.2
